Move topology to top-level public package
Part of #23 — makes the graph library importable:
import "github.com/Darkroom4364/netlens/topology"

Mechanical rename: internal/topology → topology/, updated 20 import paths.
Fixed testdata path.
