+++ internal/cli/benchmark.go
@@ -9,7 +9,7 @@ import (
 	"github.com/Darkroom4364/netlens/internal/bench"
 	"github.com/Darkroom4364/netlens/internal/measure"
 	"github.com/Darkroom4364/netlens/tomo"
-	"github.com/Darkroom4364/netlens/internal/topology"
+	"github.com/Darkroom4364/netlens/topology"
 	"github.com/spf13/cobra"
 )
