+++ internal/measure/simulate.go
@@ -5,7 +5,7 @@ import (
 	"math/rand"
 
 	"github.com/Darkroom4364/netlens/tomo"
-	"github.com/Darkroom4364/netlens/internal/topology"
+	"github.com/Darkroom4364/netlens/topology"
 )
 
 // SimConfig controls synthetic measurement generation.
