+++ internal/plan/design_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/Darkroom4364/netlens/tomo"
-	"github.com/Darkroom4364/netlens/internal/topology"
+	"github.com/Darkroom4364/netlens/topology"
 	"gonum.org/v1/gonum/mat"
 )
