+++ tomo/conformal_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/Darkroom4364/netlens/tomo"
-	"github.com/Darkroom4364/netlens/internal/topology"
+	"github.com/Darkroom4364/netlens/topology"
 )
 
 func TestConformalTriangleNoiseFree(t *testing.T) {
