+++ topology/edge_test.go
