+++ topology/fuzz_test.go
