+++ topology/infer.go
